refactor(web): extract EventTicker formatter + vitest behavioral coverage
Builds on leno23's args.newLevel fix (closes #173) with three follow-ups
gemini-code-assist flagged:

- Extract `formatChainEvent` + helpers to apps/web/src/eventTickerFormat.ts
  (React-free pure module) so the formatter can be unit-tested without
  pulling in React, Convex client, or browser-only modules.
- Delete fragile readFileSync+regex test in packages/shared/test/
  clanWorldAbi.test.ts; the ABI-shape pin stays, but the behavioral check
  ("ticker reads args.newLevel") moves to a real test against the pure
  formatter where renames/regressions surface as failed assertions, not a
  brittle source-file regex mismatch.
- Add vitest config to apps/web (was Playwright-only) + replace the dead
  apps/web/src/EventTicker.test.ts (used @playwright/test but lived
  outside the playwright testDir, so it never ran) with
  apps/web/src/eventTickerFormat.test.ts: 6 cases covering WorldPaused,
  WorldUnpaused, ClansmanRevived, WallDamagedByBandit happy/legacy paths,
  and unknown event fallback.

Results:
- pnpm --filter @clan-world/web test → 6/6 pass
- pnpm --filter @clan-world/shared test → 31/31 pass (was 32; fragile
  test removed)
- pnpm --filter @clan-world/web typecheck → clean
