+++ apps/web/package.json
@@ -9,6 +9,7 @@
     "preview": "vite preview",
     "typecheck": "tsc --noEmit",
     "lint": "echo 'lint stub'",
+    "test": "vitest run",
     "clean": "rm -rf dist .turbo node_modules/.vite",
     "playwright": "playwright"
   },
@@ -29,6 +30,7 @@
     "@types/react-dom": "18.3.1",
     "@vitejs/plugin-react": "4.3.4",
     "typescript": "5.9.3",
-    "vite": "5.4.11"
+    "vite": "5.4.11",
+    "vitest": "^3.2.0"
   }
 }
